test(triggers): add team-requested case and extract shared test helper
Address PR review items 1 and 2: add explicit test verifying
review_requested events targeting teams (not individuals) result in
bot_not_requested skip, and extract duplicate createMockGitHubContext
into triggers/test-helpers.ts to eliminate copy-paste across test files.
