+++ src/features/triggers/router.test.ts
@@ -1,27 +1,11 @@
-import type {GitHubContext} from '../../services/github/types.js'
 import type {Logger} from '../../shared/logger.js'
 import type {TriggerConfig} from './types.js'
 import {beforeEach, describe, expect, it} from 'vitest'
-import {classifyEventType, normalizeEvent} from '../../services/github/context.js'
 import {createMockLogger} from '../../shared/test-helpers.js'
 import {createIssueCommentCreatedEvent} from './__fixtures__/payloads.js'
 import {checkSkipConditions, extractCommand, hasBotMention, routeEvent} from './router.js'
+import {createMockGitHubContext} from './test-helpers.js'
 import {ALL_AUTHOR_ASSOCIATIONS, ALLOWED_ASSOCIATIONS} from './types.js'
-
-function createMockGitHubContext(eventName: string, payload: unknown = {}): GitHubContext {
-  const eventType = classifyEventType(eventName)
-  return {
-    eventName,
-    eventType,
-    repo: {owner: 'owner', repo: 'repo'},
-    ref: 'refs/heads/main',
-    sha: 'abc123',
-    runId: 12345,
-    actor: 'actor',
-    payload,
-    event: normalizeEvent(eventType, payload),
-  }
-}
 
 describe('hasBotMention', () => {
   it('detects @botname mention', () => {
